+++ src/extensions/pxblue-extensions.ts
@@ -15,6 +15,7 @@ import {
     PRETTIER_DEPENDENCIES,
     PRETTIER_SCRIPTS,
     PRETTIER_CONFIG,
+    APP_COMPONENT,
 } from '../constants';
 import {
     updateScripts,
@@ -131,9 +132,9 @@ module.exports = (toolbox: GluegunToolbox): void => {
         filesystem.write(`${folder}/package.json`, packageJSON, { jsonIndent: 4 });
 
         // Update browsers list
-        let browsers = filesystem.read(`${folder}/browserslist`, 'utf8');
+        let browsers = filesystem.read(`${folder}/.browserslistrc`, 'utf8');
         browsers = updateBrowsersListFile(browsers);
-        filesystem.write(`${folder}/browserslist`, browsers);
+        filesystem.write(`${folder}/.browserslistrc`, browsers);
 
         // Update index.html
         let html = filesystem.read(`${folder}/src/index.html`, 'utf8');
@@ -210,9 +211,9 @@ module.exports = (toolbox: GluegunToolbox): void => {
                 config: LINT_CONFIG.tsx,
             });
 
-            let serviceWorker = filesystem.read(`${folder}/src/serviceWorker.ts`);
-            serviceWorker = `/* eslint-disable */\r\n${serviceWorker}`;
-            filesystem.write(`${folder}/src/serviceWorker.ts`, serviceWorker);
+            let webVitals = filesystem.read(`${folder}/src/reportWebVitals.ts`);
+            webVitals = `/* eslint-disable */\r\n${webVitals}`;
+            filesystem.write(`${folder}/src/reportWebVitals.ts`, webVitals);
         }
 
         // Install Code Formatting Packages (optional)
@@ -244,9 +245,16 @@ module.exports = (toolbox: GluegunToolbox): void => {
         let index = filesystem.read(`${folder}/src/index.${!ts ? 'js' : 'tsx'}`, 'utf8');
         const imports = ROOT_IMPORTS.react.join('\r\n');
         index = `import 'react-app-polyfill/ie11';\r\nimport 'react-app-polyfill/stable';\r\n${index}`
-            .replace("'./serviceWorker';", `'./serviceWorker';\r\n${imports}\r\n`)
+            .replace('ReactDOM.render(', `${imports}\r\n\r\nReactDOM.render(`)
             .replace('<App />', ROOT_COMPONENT.react);
         filesystem.write(`${folder}/src/index.${!ts ? 'js' : 'tsx'}`, index);
+
+        if (ts && lint) {
+            // update the App.tsx to pass linting
+            let app = filesystem.read(`${folder}/src/App.${!ts ? 'js' : 'tsx'}`, 'utf8');
+            app = APP_COMPONENT.react;
+            filesystem.write(`${folder}/src/App.${!ts ? 'js' : 'tsx'}`, app);
+        }
 
         // Update index.html
         let html = filesystem.read(`${folder}/public/index.html`, 'utf8');